+++ web/src/pages/Graphdb/index.tsx
@@ -1,4 +1,4 @@
-import { Button, message, Popconfirm, Tag } from 'antd'
+import { message } from 'antd'
 import styles from './index.less'
 import { useImmer } from 'use-immer'
 import GraphDataModal from './components/GraphDataModal'
@@ -6,6 +6,7 @@ import { useDatabaseEntity } from '@/domains/entities/database-manager'
 import useIntlConfig from '@/hooks/useIntlConfig';
 import AsyncTable from '@/components/AsyncTable'
 import { useEffect } from 'react'
+import { useGraphDBColumns } from '@/hooks/useGraphDBColumns'
 
 const Graphdb: React.FC = () => {
   const [state, setState] = useImmer<{
@@ -23,10 +24,10 @@ const Graphdb: React.FC = () => {
     getDatabaseList()
   };
 
-  const onOpenModal = (id = null) => {
+  const onOpenModal = (id?: string) => {
     setState((draft) => {
       draft.open = true
-      draft.editId = id
+      draft.editId = id || null;
     })
   }
 
@@ -61,55 +62,15 @@ const Graphdb: React.FC = () => {
   }, [])
 
 
-  const columns = [
-    {
-      title: formatMessage('database.columns.name'),
-      dataIndex: 'name',
-      render: (text: string, record: any) => {
-        return <div className={styles['graph-database-name']}>
-          {text}
-          {record.is_default_db && <Tag style={{ marginLeft: 10 }} bordered={false} color="processing">
-            {formatMessage('database.columns.defaultTag')}
-          </Tag>}
-        </div>
-      }
-    },
-    {
-      title: formatMessage('database.columns.ip'),
-      dataIndex: 'ip',
-    },
-    {
-      title: formatMessage('database.columns.default'),
-      dataIndex: 'desc',
-    },
-    {
-      title: formatMessage('database.columns.status'),
-      dataIndex: 'stauts',
-      render: (text: boolean) => {
-        return text ? <Tag color="cyan" bordered={false}>可用</Tag> : <Tag color="error" bordered={false}>不可用</Tag>
-      }
-    },
-    {
-      title: formatMessage('database.columns.updateTime'),
-      dataIndex: 'updateTime',
-    },
-    {
-      title: formatMessage('database.columns.operation'),
-      dataIndex: 'operation',
-      render: (text: string, record: any) => {
-        return <div className={styles['graph-database-operation']}>
-          <Button type="link" onClick={() => onOpenModal(record.id)} >{formatMessage('actions.edit')}</Button>
-          <Popconfirm
-            title={formatMessage('database.deleteConfirm', { name: record.name })}
-            onConfirm={() => onDeleteGraphDatabase(record.id)}
-          >
-            <Button type="link" disabled={record.is_default_db}>{formatMessage('actions.delete')}</Button>
-          </Popconfirm>
-          <Button type="link" disabled={record.is_default_db} onClick={() => setDefaultGraphDatabase(record)}>{formatMessage('actions.setDefault')}</Button>
-        </div>
-      }
-    },
-  ]
+  const { columns } = useGraphDBColumns({
+    styles,
+    onDeleteGraphDatabase,
+    setDefaultGraphDatabase,
+    onOpenModal,
+  })
+
+
+
 
   return <div className={styles['graph-database']}>
     <div className={styles['graph-database-title']}>{formatMessage('database.title')}</div>
