+++ web/src/pages/Home/index.less
@@ -223,3 +223,14 @@
     }
   }
 }
+
+
+.chat-emty>.messages {
+  :global {
+    .ant-bubble-start .ant-bubble-content {
+      border-color: transparent;
+      border-radius: 0;
+      overflow: visible;
+    }
+  }
+}
