+++ web/src/pages/Home/index.tsx
@@ -422,7 +422,9 @@ const HomePage: React.FC = () => {
         }
       </div>
 
-      <div className={styles.chat}>
+      <div className={
+        [styles.chat,
+        !items?.length ? styles['chat-emty'] : ''].join(' ')}>
         {/* 消息列表 */}
         <Bubble.List
           items={items.length > 0 ? items : [{
